Test configuration of project.
