+++ tests/ConfigurationException/InvalidConfigurationExceptionTest.php
@@ -12,7 +12,7 @@
 
 namespace PhpCsFixer\Tests\ConfigurationException;
 
-use PhpCsFixer\ConfigurationException\InvalidFixerConfigurationException;
+use PhpCsFixer\ConfigurationException\InvalidConfigurationException;
 use PhpCsFixer\Console\Command\FixCommand;
 use PHPUnit\Framework\TestCase;
 
@@ -21,49 +21,42 @@ use PHPUnit\Framework\TestCase;
  *
  * @internal
  *
- * @covers \PhpCsFixer\ConfigurationException\InvalidFixerConfigurationException
+ * @covers \PhpCsFixer\ConfigurationException\InvalidConfigurationException
  */
 final class InvalidConfigurationExceptionTest extends TestCase
 {
-    public function testIsInvalidConfigurationException()
+    public function testIsInvalidArgumentException()
     {
-        $exception = new InvalidFixerConfigurationException(
-            'hal',
-            'I cannot do that, Dave.'
-        );
+        $exception = new InvalidConfigurationException('I cannot do that, Dave.');
 
-        $this->assertInstanceOf('PhpCsFixer\ConfigurationException\InvalidConfigurationException', $exception);
+        $this->assertInstanceOf('InvalidArgumentException', $exception);
     }
 
     public function testDefaults()
     {
-        $fixerName = 'hal';
         $message = 'I cannot do that, Dave.';
 
-        $exception = new InvalidFixerConfigurationException(
-            $fixerName,
-            $message
-        );
+        $exception = new InvalidConfigurationException($message);
 
-        $this->assertSame(sprintf('[%s] %s', $fixerName, $message), $exception->getMessage());
-        $this->assertSame(FixCommand::EXIT_STATUS_FLAG_HAS_INVALID_FIXER_CONFIG, $exception->getCode());
+        $this->assertSame($message, $exception->getMessage());
+        $this->assertSame(FixCommand::EXIT_STATUS_FLAG_HAS_INVALID_CONFIG, $exception->getCode());
         $this->assertNull($exception->getPrevious());
     }
 
     public function testConstructorSetsValues()
     {
-        $fixerName = 'hal';
         $message = 'I cannot do that, Dave.';
+        $code = 9000;
         $previous = new \RuntimeException();
 
-        $exception = new InvalidFixerConfigurationException(
-            $fixerName,
+        $exception = new InvalidConfigurationException(
             $message,
+            $code,
             $previous
         );
 
-        $this->assertSame(sprintf('[%s] %s', $fixerName, $message), $exception->getMessage());
-        $this->assertSame(FixCommand::EXIT_STATUS_FLAG_HAS_INVALID_FIXER_CONFIG, $exception->getCode());
+        $this->assertSame($message, $exception->getMessage());
+        $this->assertSame($code, $exception->getCode());
         $this->assertSame($previous, $exception->getPrevious());
     }
 }
